Fixing SetSmoothingGroup trying to parse a long as an int

+++ Source/HelixToolkit.Wpf/Importers/ObjReader.cs
@@ -421,8 +421,8 @@ namespace HelixToolkit.Wpf
             }
             else
             {
-                int smoothingGroup;
-                if (int.TryParse(values, out smoothingGroup))
+                long smoothingGroup;
+                if (long.TryParse(values, out smoothingGroup))
                 {
                     this.currentSmoothingGroup = smoothingGroup;
                 }
